backend/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.28.2
six==1.16.0
sseclient-py==1.7.2
urllib3==1.26.16
Werkzeug==2.3.6
python-dotenv==1.0.0
tiktoken==0.5.1
regex==2023.8.8 
openai==0.28.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3 
attrs==23.1.0 
frozenlist==1.4.0 
multidict==6.0.4  
tqdm==4.66.1 
yarl==1.9.2
langchain==0.0.302
PyYAML==6.0.1
SQLAlchemy==2.0.21 
annotated-types==0.5.0 
anyio==3.7.1 
dataclasses-json==0.6.0 
greenlet==2.0.2 
jsonpatch==1.33 
jsonpointer==2.4 
langsmith==0.0.40 
marshmallow==3.20.1 
mypy-extensions==1.0.0 
numexpr==2.8.6 
numpy==1.26.0 
packaging==23.1 
pydantic==2.4.1 
pydantic-core==2.10.1 
sniffio==1.3.0 
tenacity==8.2.3 
typing-extensions==4.8.0 
typing-inspect==0.9.0
pinecone-client==2.2.4
dnspython==2.4.2 
loguru==0.7.2  
win32-setctime==1.1.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2 
bs4==0.0.1 
soupsieve==2.5
elastic-site-search==2.1.1
lxml==4.9.3